+++ src/entity.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "common.h"
 #include "raylib.h"
 #include "raymath.h"
 #include "window.h"
@@ -50,13 +51,13 @@ RayCollision GetMouseRayCollisionBase(Base base, Camera camera);
 
 EntityGroup* CreateEntityGroup();
 
-void AddGroupEntity(EntityGroup* group, void* data, size_t size, Components comps,
+void AddGroupEntity(EntityGroup* group, void* data, uint size, Components comps,
                     UPDATE_FUNC updateFunc, DRAW_FUNC drawFunc);
 
 bool CheckEntityComponents(Entity* entity, Components filter); // Checks if an entity has a tag/component
 
-size_t PollEntities(EntityGroup* group, Components filter, POLL_FUNC func);
+uint PollEntities(EntityGroup* group, Components filter, POLL_FUNC func);
 
-size_t UpdateGroup(EntityGroup* group, float delta);
+uint UpdateGroup(EntityGroup* group, float delta);
 
-size_t DrawGroup(EntityGroup* group);
+uint DrawGroup(EntityGroup* group);
